+++ src/identity_map/iter/mod.rs
@@ -81,7 +81,7 @@ impl<K, V> Default for Iter<'_, K, V> {
 impl<K, V> DoubleEndedIterator for Iter<'_, K, V> {
 	#[inline(always)]
 	fn next_back(&mut self) -> Option<Self::Item> {
-		self.iter.next_back()
+		self.iter.next_back().map(|(k, v)| (k, v))
 	}
 }
 
@@ -90,11 +90,11 @@ impl<K, V> ExactSizeIterator for Iter<'_, K, V> { }
 impl<K, V> FusedIterator for Iter<'_, K, V> { }
 
 impl<'a, K, V> Iterator for Iter<'a, K, V> {
-	type Item = &'a (K, V);
+	type Item = (&'a K, &'a V);
 
 	#[inline(always)]
 	fn next(&mut self) -> Option<Self::Item> {
-		self.iter.next()
+		self.iter.next().map(|(k, v)| (k, v))
 	}
 
 	#[inline(always)]